Make CGRect editor strongly typed.
