+++ Xamarin.PropertyEditing.Mac/Xamarin.PropertyEditing.Mac.csproj
@@ -66,7 +66,8 @@
     <Compile Include="Controls\EnumEditorControl.cs" />
     <Compile Include="Controls\ConstraintExtensions.cs" />
     <Compile Include="Controls\BaseRectangleEditorControl.cs" />
-    <Compile Include="Controls\RectangleEditorControl.cs" />
+    <Compile Include="Controls\CGRectEditorControl.cs" />
+    <Compile Include="Controls\PointEditorControl.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Controls\" />
